requirements.txt
python-telegram-bot==13.7
yt-dlp==2021.12.01
shazamio==1.0.2
pydub==0.25.1
requests==2.25.1
